requirements.txt
selenium
selenium-stealth
python-dotenv
tqdm
beautifulsoup4
requests